[DESIGN] : 🎨 유저 생성, 조회 페이지 Css, Ui 구현

+++ src/pages/Business/Business.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import styled from 'styled-components';
 import MaxInput from '../../components/Inputs/Input/MaxInput';
 import ButtonInput from '../../components/Inputs/ButtonInput/ButtonInput';
 import Dropdown from '../../components/Dropdown/Dropdown';
@@ -11,11 +10,12 @@ import 'react-datepicker/dist/react-datepicker.css';
 import { useUserIdValidation } from './hooks/useUserIdValidation';
 import { useSignup } from './hooks/useSignup';
 import './business.css';
-import { ko } from 'date-fns/locale';
 import { Users, useGetUser } from './hooks/useGetUser';
 import { useDeleteUser } from './hooks/useDeletUser';
-import DropdownPortal from '../../components/Dropdown/DropdownPortal';
 import { usePatchUser } from './hooks/usePatchUser';
+import { Wrapper, ViewUser, VuHeader, Vubody, StSpan, UserInfo } from './styles';
+import CreateUser from './CreateUser';
+import Swal from 'sweetalert2';
 
 const Business = () => {
   // 모달 상태변수, 콜백함수
@@ -23,69 +23,6 @@ const Business = () => {
 
   const closeModal = () => {
     setShowModal(false);
-  };
-
-  // 유저생성 상태변수
-  const [userInfo, setUserInfo] = React.useState<UserSignupInfo>({
-    userName: '',
-    team: '',
-    rank: '',
-    job: '',
-    userId: '',
-    joinDay: new Date(),
-    salaryDay: '',
-    authLevel: 0,
-  });
-
-  // 유저생성 인풋 체인지 핸들러
-  const changeInputHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    const updateUserInfo = name === 'salaryDay' ? Number(value) : value;
-    setUserInfo({ ...userInfo, [name]: updateUserInfo });
-  };
-
-  // 유저생성 권한 드롭 다운 체인지 핸들러
-  const selecteAuthorityHandler = (value: number | string) => {
-    setUserInfo({ ...userInfo, authLevel: Number(value) });
-  };
-
-  const selecteAuthorityPatchHandler = (value: number | string) => {
-    setPatchUserInfo({ ...patchUserInfo, authLevel: Number(value) });
-  };
-
-  // datepicker handler
-  const changeDateHandler = (date: Date) => {
-    setUserInfo({ ...userInfo, joinDay: date });
-  };
-
-  // 권한 드롭 다운 배열
-  const authority = [
-    { name: '관리자', value: 2 },
-    { name: '직원', value: 3 },
-  ];
-  // 아이디 유효성 검사, 중복확인 콜백함수
-  const { validUserId, checkUserId, userIdValidation, setUserIdValidation } =
-    useUserIdValidation();
-
-  // 아이디 유호성 검사, 중복확인 핸들러
-  const checkUserIdHandler = (item: string) => {
-    if (validUserId(item)) {
-      checkUserId.mutate(item);
-    } else {
-      setUserIdValidation(false);
-      alert('아이디 양식을 재확인 해주세요');
-    }
-  };
-
-  const { signup } = useSignup();
-
-  const submitSignInfoHandler = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (userIdValidation) {
-      signup.mutate(userInfo);
-    } else {
-      alert('가입에 실패하였습니다 입력한 내용을 확인해주세요');
-    }
   };
 
   const { data, isLoading } = useGetUser();
@@ -101,6 +38,10 @@ const Business = () => {
     authLevel: '',
   });
 
+  const selecteAuthorityPatchHandler = (value: number | string) => {
+    setPatchUserInfo({ ...patchUserInfo, authLevel: Number(value) });
+  };
+
   const handleUserClick = (user: Users) => {
     setSelectedUser(user);
     setShowModal(true);
@@ -114,8 +55,24 @@ const Business = () => {
   const { deleteUser } = useDeleteUser();
 
   const deleteUserHandler = (): void => {
-    deleteUser(selectedUser.userId);
     closeModal();
+
+    Swal.fire({
+      title: '정말 삭제하시겠습니까?',
+      text: '삭제된 유저는 복구되지 않습니다.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '삭제',
+    }).then(result => {
+      if (result.isConfirmed) {
+        Swal.fire({ title: '정상적으로 삭제 되었습니다.', icon: 'success' });
+        deleteUser(selectedUser.userId);
+      } else {
+        setShowModal(true);
+      }
+    });
   };
 
   interface PatchUserInfo {
@@ -149,227 +106,300 @@ const Business = () => {
   const { patchUser } = usePatchUser();
 
   const patchUserHandler = () => {
-    patchUser(selectedUser.userId, patchUserInfo);
-    alert('수정되었습니다.');
     closeModal();
+
+    Swal.fire({
+      title: '유저 정보를 수정 하시겠습니까?',
+      text: '수정된 정보는 되돌릴 수 없습니다.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '수정',
+    }).then(result => {
+      if (result.isConfirmed) {
+        Swal.fire({ title: '정상적으로 수정 되었습니다.', icon: 'success' });
+        patchUser(selectedUser.userId, patchUserInfo);
+      } else {
+        setShowModal(true);
+      }
+    });
   };
 
+  // 권한 드롭 다운 배열
+  const authority = [
+    { name: '관리자', value: 2 },
+    { name: '직원', value: 3 },
+  ];
+
   return (
-    <MainWrapper>
+    <Wrapper>
       <ViewUser>
         <VuHeader>
+          <StSpan>유저 조회</StSpan>
+          <StSpan style={{ margin: '0 91px 0 60px' }}>부서별 보기</StSpan>
           {/* <Dropdown items={teams}>부서</Dropdown> */}
-          <MaxInput types="max" type="search">
-            검색
-          </MaxInput>
+          <MaxInput
+            types="login"
+            type="search"
+            style={{
+              width: '360px',
+              height: '33px',
+              background: 'none',
+              boxShadow: 'none',
+              borderBottom: '1px solid black',
+              boxSizing: 'border-box',
+              padding: '0 15px 15px',
+            }}
+          />
         </VuHeader>
         <Vubody>
+          <div
+            style={{
+              borderBottom: '1px solid #484240',
+              padding: '0 20px',
+              boxSizing: 'border-box',
+              display: 'flex',
+              justifyContent: 'space-between',
+              fontSize: '12px',
+              gap: '80px',
+            }}
+          >
+            <div
+              style={{
+                display: 'flex',
+                flexDirection: 'row',
+                gap: '50px',
+              }}
+            >
+              <UserInfo>부서</UserInfo>
+              <UserInfo>직급</UserInfo>
+            </div>
+            <div
+              style={{
+                display: 'flex',
+                flexDirection: 'row',
+                gap: '50px',
+              }}
+            >
+              <UserInfo>직무</UserInfo>
+              <UserInfo>이름</UserInfo>
+              <UserInfo>입사 일자</UserInfo>
+            </div>
+          </div>
           {data?.map(user => {
             return (
-              <div key={user.userId} onClick={() => handleUserClick(user)}>
-                <span>{user.team}</span>&nbsp;
-                <span>{user.rank}</span>&nbsp;
-                <span>{user.job}</span>&nbsp;
-                <span>{user.userName}</span>&nbsp;
-                <span>{String(user.joinDay)}</span>&nbsp;
+              <div
+                key={user.userId}
+                onClick={() => handleUserClick(user)}
+                style={{
+                  padding: '0 20px',
+                  display: 'flex',
+                  justifyContent: 'space-between',
+                  gap: '80px',
+                }}
+              >
+                <div
+                  style={{
+                    borderBottom: '1px solid #484240',
+                    display: 'flex',
+                    flexDirection: 'row',
+                    gap: '50px',
+                  }}
+                >
+                  <UserInfo>{user.team}</UserInfo>
+                  <UserInfo>{user.rank}</UserInfo>
+                </div>
+                <div
+                  style={{
+                    borderBottom: '1px solid #484240',
+                    display: 'flex',
+                    flexDirection: 'row',
+                    gap: '50px',
+                  }}
+                >
+                  <UserInfo>{user.job}</UserInfo>
+                  <UserInfo>{user.userName}</UserInfo>
+                  <UserInfo>{String(user.joinDay)}</UserInfo>
+                </div>
               </div>
             );
           })}
           {showModal && (
-            <Modal closeModal={closeModal}>
-              이름{'  '}
-              {selectedUser.userName}
+            <Modal closeModal={closeModal} style={{ width: '640px' }}>
+              <h1
+                style={{
+                  width: '500px',
+                  fontSize: '18px',
+                  fontWeight: 'bolder',
+                  borderBottom: '1px solid black',
+                  paddingBottom: '15px',
+                  boxSizing: 'border-box',
+                  marginBottom: '25px',
+                  textAlign: 'center',
+                  margin: '30px 0 30px',
+                }}
+              >
+                유저 조회
+              </h1>
               <MaxInput
-                types="max"
+                types="signup"
+                style={{ width: '500px' }}
+                value={selectedUser.userName}
+              >
+                이름
+              </MaxInput>
+              <MaxInput
+                types="signup"
+                style={{ width: '500px', color: '#8AB2E0' }}
                 value={patchUserInfo.team}
                 name="team"
                 onChange={patchInputHandler}
               >
                 부서
               </MaxInput>
-              <MaxInput
-                types="max"
-                value={patchUserInfo.rank}
-                name="rank"
-                onChange={patchInputHandler}
+              <div
+                style={{
+                  display: 'flex',
+                  flexDirection: 'row',
+                  gap: '20px',
+                  color: '#8AB2E0',
+                }}
               >
-                직급
-              </MaxInput>
-              <MaxInput
-                types="max"
-                value={patchUserInfo.job}
-                name="job"
-                onChange={patchInputHandler}
-              >
-                직무
-              </MaxInput>
-              아이디{'  '}
-              {selectedUser.userId}
-              <br />
-              입사일{'  '}
-              {String(selectedUser.joinDay)}
-              <br />
-              월급일{'  '}
-              {selectedUser.salaryDay}
-              <br />
-              권한{' : '}
-              {selectedUser.authLevel}
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px', color: '#8AB2E0' }}
+                  type="text"
+                  name="rank"
+                  value={patchUserInfo.rank}
+                  onChange={patchInputHandler}
+                >
+                  직급
+                </MaxInput>
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px', color: '#8AB2E0' }}
+                  type="text"
+                  value={patchUserInfo.job}
+                  name="job"
+                  onChange={patchInputHandler}
+                >
+                  직무
+                </MaxInput>
+              </div>
+              <div style={{ display: 'flex', flexDirection: 'row', gap: '20px' }}>
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px' }}
+                  value={String(selectedUser.joinDay)}
+                >
+                  입사일
+                </MaxInput>
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px' }}
+                  value={selectedUser.salaryDay}
+                >
+                  월급일
+                </MaxInput>
+              </div>
+              <div style={{ display: 'flex', flexDirection: 'row', gap: '20px' }}>
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px' }}
+                  value={selectedUser.userId}
+                >
+                  아이디
+                </MaxInput>
+                <MaxInput
+                  types="signup"
+                  style={{ width: '240px' }}
+                  value={selectedUser.authLevel}
+                >
+                  권한
+                </MaxInput>
+              </div>
               <Dropdown
                 size="small"
                 items={authority}
                 value={selectedUser.authLevel}
                 onChange={selecteAuthorityPatchHandler}
+                style={{
+                  width: '500px',
+                  height: '50px',
+                  boxShadow: '0 4px 4px rgba(201, 201, 201, 0.25)',
+                  fontSize: '15px',
+                  border: 'none',
+                  padding: '15px',
+                  fontWeight: 'bold',
+                  color: '#484240',
+                }}
               >
-                변경
+                권한 변경
               </Dropdown>
-              <StButton onClick={patchUserHandler}>수정</StButton>
-              <StButton onClick={deleteUserHandler}>삭제</StButton>
-              <StButton onClick={closeModal}>닫기</StButton>
+              <div
+                style={{
+                  width: '500px',
+                  display: 'flex',
+                  justifyContent: 'space-between',
+                  margin: '50px 0 ',
+                }}
+              >
+                <StButton
+                  size="signup"
+                  onClick={patchUserHandler}
+                  style={{
+                    width: '150px',
+                    boxShadow: '0 4px 4px rgba(201, 201, 201, 0.25)',
+                    background: '#E64042',
+                    color: '#fff',
+                    borderRadius: '7px',
+                    fontSize: '15px',
+                    fontWeight: 'bold',
+                  }}
+                >
+                  수정
+                </StButton>
+                <StButton
+                  size="signup"
+                  onClick={deleteUserHandler}
+                  style={{
+                    width: '150px',
+                    boxShadow: '0 4px 4px rgba(201, 201, 201, 0.25)',
+                    background: '#E64042',
+                    color: '#fff',
+                    borderRadius: '7px',
+                    fontSize: '15px',
+                    fontWeight: 'bold',
+                  }}
+                >
+                  삭제
+                </StButton>
+                <StButton
+                  size="signup"
+                  onClick={closeModal}
+                  style={{
+                    width: '150px',
+                    boxShadow: '0 4px 4px rgba(201, 201, 201, 0.25)',
+                    background: '#fff',
+                    color: '#E64042',
+                    border: '1px solid #E64042',
+                    boxSizing: 'border-box',
+                    borderRadius: '7px',
+                    fontSize: '15px',
+                    fontWeight: 'bold',
+                  }}
+                >
+                  닫기
+                </StButton>
+              </div>
             </Modal>
           )}
         </Vubody>
       </ViewUser>
-      <CreateUser onSubmit={submitSignInfoHandler}>
-        <h1>유저 생성</h1>
-        {/* <-----------이름 : userName-----------> */}
-        <MaxInput
-          types="max"
-          type="text"
-          name="userName"
-          value={userInfo.userName}
-          onChange={changeInputHandler}
-        >
-          이름
-        </MaxInput>
-        {/* <-----------이름 : userName-----------> */}
-        {/* <-----------부서 : team-----------> */}
-        <MaxInput
-          types="max"
-          type="text"
-          name="team"
-          value={userInfo.team}
-          onChange={changeInputHandler}
-        >
-          부서
-        </MaxInput>
-        {/* <-----------부서 : team----------> */}
-        {/* <-----------직급 : rank-----------> */}
-        <MaxInput
-          types="max"
-          type="text"
-          name="rank"
-          value={userInfo.rank}
-          onChange={changeInputHandler}
-        >
-          직급
-        </MaxInput>
-        {/* <-----------직급 : rank-----------> */}
-        {/* <-----------직무 : job-----------> */}
-        <MaxInput
-          types="max"
-          type="text"
-          name="job"
-          value={userInfo.job}
-          onChange={changeInputHandler}
-        >
-          직무
-        </MaxInput>
-        {/* <-----------직무 : job-----------> */}
-        {/* <-----------아이디 : userID-----------> */}
-        <ButtonInput
-          types="button"
-          type="text"
-          name="userId"
-          value={userInfo.userId}
-          onChange={changeInputHandler}
-          onClick={() => checkUserIdHandler(userInfo.userId)}
-          buttonTag="중복검사"
-        >
-          아이디
-        </ButtonInput>
-        {/* <-----------아이디 : userID-----------> */}
-        {/* <-----------입사일 : joinDay-----------> */}
-        입사일
-        <br />
-        <DatePicker
-          selected={userInfo.joinDay}
-          onChange={changeDateHandler}
-          locale={ko}
-          dateFormat="yyyy년 MM월 dd일"
-        />
-        {/* <-----------입사일 : joinDay-----------> */}
-        {/* <-----------월급일 : salaryDay-----------> */}
-        <br />
-        <MaxInput
-          types="max"
-          type="text"
-          name="salaryDay"
-          value={userInfo.salaryDay}
-          onChange={changeInputHandler}
-        >
-          월급일
-        </MaxInput>
-        {/* <-----------월급일 : salaryDay-----------> */}
-        {/* <-----------권한 : authLevel-----------> */}
-        <br />
-        <Dropdown
-          size="small"
-          items={authority}
-          value={userInfo.authLevel}
-          onChange={selecteAuthorityHandler}
-        >
-          권한
-        </Dropdown>
-        {/* <-----------권한 : authLevel-----------> */}
-        <br />
-        <StButton>생성</StButton>
-      </CreateUser>
-    </MainWrapper>
+      <CreateUser />
+    </Wrapper>
   );
 };
 
 export default Business;
-
-const MainWrapper = styled.div`
-  width: 1200px;
-  height: 800px;
-  margin: auto;
-
-  display: flex;
-  flex-direction: row;
-
-  border: 1px solid black;
-  box-sizing: border-box;
-`;
-
-const ViewUser = styled.div`
-  width: 600px;
-  height: 100%;
-
-  border: 1px solid red;
-  box-sizing: border-box;
-`;
-
-const CreateUser = styled.form`
-  width: 600px;
-  height: 100%;
-
-  border: 1px solid blue;
-  box-sizing: border-box;
-`;
-
-const VuHeader = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-const Vubody = styled.div`
-  width: 500px;
-  height: 600px;
-  margin: 20px auto;
-  padding: 20px;
-
-  border: 1px solid black;
-  box-sizing: border-box;
-`;
